Questions from database to a candidate view
I made a jsp-file (based on Jennas html one) and made a servlet for bringing questions from a database to a .jsp-servlet. Update Dao-file by making a new method, which reads data from database to on arraylist. Data-file Questions.java for to a transfer data infos. Finally add this new servlet to an index-page.